# Improve `take` kernel performance on primitive arrays, fix bad null index handling  (#4404)